bot/requirements.txt
# TODO versionize
# Framework
flask
gunicorn
sentry-sdk[flask]
supervisor
aiohttp
requests

# Discord
unidecode
discord.py
